Simplify return type of `neighbors_of`
It seems weird to not signal the "vertex not in polygon" case with a
specific return value, but I'm currently working on updating this method
to use `self.0.edges` and all the `Option` does is complicate the
implementation.

If a vertex has no neighbors, then it not being in the polygon is the
only reason for that though (or a bug in modifying the polygon,
possibly). So strictly speaking, the new return value is just as
expressive. It just seems weird.