Simplify signature of unimplemented method
The old signature assumed that, conceptually, the grid is bounded by its
outer edges. This made everything complicated, as cube center neighbors
can still be easily computed for these outer edges.

Let's just say that, conceptually, the grid is bounded by a layer of
outer cubes, outside of the outermost edges. That removes one edge case
that would otherwise need to be considered.