Rename `Transform` to `Camera`
I've kinda resisted the name "camera", because the metaphor we're using
is not that of a camera freely flying through a static scene, but of a
static camera that looks at a movable and rotateable model.

But I have to admit, it's the more fitting name. The metaphor can be
documented, to prevent confusion.