Use `assert_relative_eq` to test `Plane` equality
This is part of an attempt to fix #47. I don't think it will actually
help, but given that I don't fully understand what's going on in the
first place, I just want to see what happens.

But regardless of that, `relative_eq` is the better way to make these
kinds of comparisons anyway.