Fix navigation links
This is pretty annoying, as the links to "About" no longer works in
local preview mode now. But at least this should patch things up so they
work in production.

This could be fixed with a custom domain name, which I didn't want to
look into this early, but I guess the priority for that task just jumped
up.