Remove "enclosure" model
This is a work-in-progress model that I'm using as my guide and testbed
for working on CSG features. Or rather, so far, the refactoring and
infrastructure-building required to support CSG.

As this whole project is becoming less hobbyist/experimental, I no
longer want to have an incomplete model among the examples, so I'm
removing it. I'm going to open a draft PR right after to re-add it. This
PR will stay as a draft until the model is fully working. It will also
serve as my WIP development branch for the CSG work.