Define line by two points
This reverts a decision that was made earlier, to define the line by a
point and a vector instead. This was probably a mistake; at least it
seems that way in the context of #78.

At the very least, this will make the implementation of #78 simpler, as
the two points can just be cached, and there doesn't need to be a
special mechanism to cache the result of computation instead (for now).