Remove workaround in triangulation code
I'm not sure what made this workaround unnecessary, but one of the
recent clean-ups must have done it. All I know is, I can't tell why it
would be necessary with the current code, and the star model, which
previously exposed the bug this workaround fixes (#60) works without any
problems now.