Rename struct field
I at first preferred `point`, due to the consistency that provides with
`Edge` (`curve`) and `Face` (`surface`). But I've come to the
conclusion, that consistency within the `Vertex` code is more important.

It also makes sense. The location of an edge is defined by its curve,
and the bounding vertices on the curve. The location of a face is
defined by its surface, and the bounding edges on the surface. It just
so happens that vertices are fully defined by their point only, without
any additional bounding elements.