Define circle radius using 2D vector
The 3D definition is kind of wrong. I think all code currently assumes
that circles are in the x-y place[1]. Even if that weren't the case, a
single 3D vector would not be enough to define the plane of the circle.

I think the best solution for now is to make this change, and
investigate how the circle representation needs to be improved later.

[1] This would mean that circles and shapes based on circles can't
    currently be rotated. There's no issue to track this, so if this is
    true, one should be opened. I've added investigating this to my task
    list.