Make edge approximation more robust
With this change, we no longer rely on curve approximations to deliver
accurate representations of the edge vertices. I add a comment in the
commit that explains why relying on this could be a problem.

This change gives the curves some more flexibility. So far, they need to
deliver these accurate representations of the vertices, which limits the
kinds of curves that can be implemented. This limitation is now lifted,
enabling the implementation of other kinds of curves, like lines that
were swept from a vertex.