Change `Line` to point-vector form
This is the more natural representation. It allowed for a slight
simplification, and the tests become a bit more readable.

It already used to be like that a while ago, but was changed to fix
floating point accuracy bugs. Since then, the approximation code had
been cleaned up and more robust, so it's possible to go back to this
representation now.