Don't use `crate::math` in `camera`
I'm about to make changes to `crate::math` that are most useful in
`kernel`. It will be tedious to adapt all code outside the kernel to
these changes, so I've decided to no longer use `crate::math` outside of
the kernel.

This decision might be reverted later, once the dust has settled a bit.
But for now, this isn't too much of a loss.