refactor: rename the binaries workflow to CD (continuous deployment)
This workflow runs on every change of the repository `main` branch and
should take care to either provide intermediary binaries, i.e. binaries
which are not supposed to be part of an official release, or to discover
that `Cargo.toml`'s version number has been changed and create an
official release on GitHub and crates.io.