Fix method being too broadly available
This implementation (transforming only the canonical form) makes sense
for 1-dimensional vertices. If the edge it bounds is transformed, then
the 1D form of the vertex doesn't need to be transformed, as the curve,
and therefore the vertices whole coordinate system, is transformed with
the edge. The canonical form needs to be transformed, so the conversion
back into 3D space is correct.

However, this method was available for 3D vertices too. In that case,
the implementation is incorrect, as the vertex location is identical to
its canonical form, and both need to be kept identical.