Add new `algorithms` module
This commit moves `approximation` and `triangulation` into a new
`kernel::algorithms` module.

Approximation and triangulation are examples of relatively self-
contained functionality, that uses the structures defined in `geometry`
and `topology` to do stuff, without those structures really needing to
know about it. This vision isn't fully realized, as especially
triangulation is entangled with `topology`, but this will be cleaned up
in due time.

It makes sense to have each of these algorithms in a central place,
where they can be understood and unit-tested, instead of spread over
lots of small methods in the `geometry` and `topology` modules. I think
we'll need to consolidate more functionality in this way:
transformations and sweep are candidates for this. Adding all of this to
the top-level namespace in the `kernel` module would become confusing.