Change the way `Vertices` works
Turn it into a temporary struct that is constructed on-demand to access
vertices or make changes to them.

The goal is for `Shape` to eventually enforce consistency through the
API it provides. This change pre-emptively plugs a hole that would allow
the caller to break this consistency, by replacing the `Vertices`
instance with another one, through the `&mut Vertices` that
`Shape::vertices` returned previously.

This also provides some more flexibility to `Shape's` implementation, as
data structures can be shared as necessary, between the different APIs,
while still allowing those APIs to be separate. This separation provides
clarity to the caller, who gets a cleaner API, but also on the
implementation side, as parts of the API can be better separated into
different modules.