Store edge vertices in 3D
This is an issue I went back and forth on a few time already.

It seems less robust, as nothing prevents these vertices from having
nothing to do at all with the curve they are supposed to be on, and I
can think of reasons why it can't stay this way long-term.

But the facts are, storing edge vertices in 1D isn't necessary right
now, makes things more complicated, and makes my ongoing work right now
a million times more difficult.

I'm pretty sure I'll want to revert this decision again some time in the
future, but for now, this is the right call.