Shade "back sides" the same as "front sides"
The shader used to shade the "back side" of a face, as defined by the
vertex normal, differently than the front sides.

This minimally invasive change removes that distinction. A surface
normal is still needed for shading, to compute the angle between the
surface and the light, but the direction of that normal is ignored.

Close #173