Fix panic with star model
Kiddo calls `Float::nan` internally, in some situations, and this case
is triggered by the star model. I didn't notice this before, as at least
some of the other models ran fine.

This commit replaces the `kiddo::KdTree` with a simple `Vec`. I don't
think this makes much of a difference for the current models. Long-term,
I think we're going to need something more advanced anyway, to support
more validations.