Remove `Cell` out of `Handle`
The `Cell` isn't needed right now, as `Handle` is only used for
immutable access.

Once mutable access is required, the `Cell` won't do. It requires its
content to be `Copy`, but `Handle` can't be `Copy` itself. Hence,
`Handle` with a `Cell` can't be used to refer to anything that itself
contains a `Handle`, which is too limiting for upcoming use cases.