Return `ValidationResult` from `add_` methods
No errors are actually returned yet. This will follow in upcoming steps.

This results in a lot of ugly `unwrap`s all over the code, but I think
that's a worthwhile trade-off. Those `unwrap`s will go away, as error
handling in the CAD kernel is cleaned up over time. The immediate
benefit is, that unit-testing the `Shape` API becomes easier.