Add bottom faces explicitly
Instead of cloning the source face, then adding the top faces, this
commit updates the sweep algorithm to add the bottom faces explicitly.
This adds a lot of tedious code. The advantage is, that now we can add
another relation of bottom-to-top topology, which is needed for creating
the side faces.

I believe that there's a better API hidden within `Shape`, that will
make code like this much cleaner. I am only starting to imagine what
that API would look like though, so for now, this mess is what we get
instead.