Move `kernel::operations` to crate root
I'm about to extract the kernel into a separate crate, and I don't want
this module to be a part of it. Nothing else in the kernel depends on
it, and it's a different thing, I think: It's the connection between the
kernel and the operations in the `fj` crate.

It's totally conceivable that someone would want to use the kernel,
without wanting to be bound to the `fj` operations. So both from a
software architecture and from a use case perspective, I think this
module should be a separate thing.