Dumb down `ValidationError`
The previous structure was pretty neat, while this is decidedly not.
What this new solution has going for it, is that it's dumb.

This is good, because `ValidationError` will soon need to be able to
represent validation errors from multiple sources, and making the
previous smart `ValidationError` event smarter, soon enough runs into
limitations of Rust's type system.

This new dumb solution, on the other hand, effortlessly scales to the
new requirements.