Rename `fj-debug` to `fj-interop`
Debug info is just one kind of data that is required for interoperation
within the Fornjot ecosystem. Triangle meshes are another example, and
there might be more in the future.

I think it's not worth having each of those in their own little crate,
so I'm designating `fj-interop` as the one crate where all such data
types can live in the future.