Don't use `mesh::Triangle` in `Face`
I only added it there very recently. It was a nice stopgap, but now it
looks like `Mesh` and `mesh::Triangle` might be developing in a
direction that makes using it there no longer appropriate.

In any case, this is actually simpler, so no downside really, regardless
of what happens with `mesh::Triangle`.