Extract export validation into separate CI job
It makes sense to have it in a separate job, so we get better feedback
on what specifically failed, if the CI build does fail because of it.

Also, I'm having trouble getting the actual 3MF validation to work on
macOS, due to linking issues. I figure that running it on Linux is good
enough for now.