Convert `Arc` into a regular circle
`Arc` doesn't follow the overall concept of the kernel: That geometry is
unbounded, and topology is bounded, referring to the unbounded geometry.

An arc, by definition, is a bounded part of a circle, not fitting into
this structure.

The reason that the former `Circle` was converted into `Arc` was to make
it possible to invert arcs/circles, which was part of fixing the surface
orientation issues. But it was a mistake to do it that way, and the new
circle representation being used now solves the problem in a more
elegant way.

This commit doesn't update the name of `Arc` yet, which is left to a
follow-up commit.