Search for version in full commit message
`release-operator` searches for the version number in a pull request's
merge commit. At the same time, it looks only at the first line of the
commit.

This is wrong, I think, as the part of the commit message I can
influence is not going to be in the first line.