Ignore 'v' prefix when searching for versions
It's optional anyway, so this should make zero difference to what the
regular expression finds. However, with the 'v' there, it's going to be
part of the match when it founds. This won't parse correctly then, using
`semver::Version::parse`, as a 'v' prefix is not a valid part of a
version.