Make `Handle` generic over `Object`
Only types that implement `Object` can be part of `Shape`. But `Handle`,
which refers to objects in `Shape`, was (up until this commit) generic
over arbitrary types. This means, it was possible to define `Handle`
types that were nonsensical, and couldn't be used in any real sense.

I did that earlier, and it would have been great, if I had noticed that
mistake right away. With the change in this commit, this mistake is no
longer possible.