Remove `distance` field of `CoherenceMismatch`
If the struct is to be used for other types of coherence mismatch, the
`distance` feel needs to be generic as well. I figure it's easier to
just remove it for now, and re-add it later as necessary, than carry it
through the coming series of changes. Especially since I have some ideas
in mind that could make dealing with local/canonical forms a bit more
straight-foward.