Remove unused update validation infrastructure
It was using the old validation infrastructure, which no longer does
anything.

With the new infrastructure, updates don't need to be validated. Either
the `Shape` they're updating is not validated, which means it needs to
be validated later anyway, or it already is, which means it can't be
updated (as `Validated` doesn't implement `DerefMut`).