Rewrite sweep algorithm
The old sweep algorithm was overly complicated and I'm pretty sure it
was also riddled with bugs that would have shown up as I continued my
cleanup work over the next weeks. It's been a thorn in my side for a
while, and while I prefer incremental improvements in general, I think a
complete rewrite was warranted in this case.

The rewritten sweep algorithm is not any shorter than the old one, but
it has a few advantages:

- It's more straight-foward than the old algorithm.
- It has a much more thorough test suite.
- It doesn't rely on any advanced or problematic features of `Shape`,
  meaning it won't block the plans to remove `Shape`.
- The needlessly complicated code that's still in there can likely be
  improved significantly once `Shape` has been removed, and the
  infrastructure that will replace it has matured a bit.