Don't check uniqueness of edges during validation
I have some changes in the works, namely replacing `Edge` with a
half-edge object, which would make this uniqueness check moot, meaning
it would have to be removed and/or adapted anyway at some point,
probably soon.

Since it's getting in the way a refactoring I'm working on, I've decided
to just remove it now.