Only pass events to input handler if `egui` doesn't want exclusive access.
This now seems to work correctly, whereas the last time I tried to
implement this functionality it didn't, IIRC.

Not sure why but might be because I'm restricting it to affecting the
"input handler"-related events only? And/or the `egui` monitored area
might be restricted to a narrower part of the window?

Anyway, this means that interacting with the UI now doesn't also
affect the model view, e.g. mouse drag can change values without
*also* rotating model. :)