Fix edge approximation ignoring vertices
The edge approximation code ignored all edge vertices when computing the
range that the curve should be approximated in. Instead, it was using
the fallback code for continuous (vertex-less) edges in every case.

This happened to work, because all the edges this approximation code was
seeing had `zero` as the local coordinate of their first vertex, which
happened to line up with the computed fallback value.

I've been making changes to the sweep algorithm (not in this branch)
that made this no longer be true, and this bug was triggered.