Add `SurfaceVertex`
Some code needs to deal with vertices in surface coordinates.
Approximation is a current example of that, but there will be more in
the future.

Calculating these surface coordinates is a bit of an error-prone
process, for the same reason that calculating global coordinates is an
error-prone process: Surface coordinates that are supposed to be the
same can be slightly different, due to floating-point accuracy.

By having a `SurfaceVertex`, we're forcing code that creates vertices to
calculate those surface coordinates once, and then they are available to
all other code that deals with these vertices. This once-at-creation
calculation can then benefit from all the protections that are available
for the construction of objects (right now, this is the validation
infrastructure, but more can be done in the future).