Remove unit test
I have to update it all the time, which is tedious, and it's nowhere
near the actually complicated code that would benefit from testing.

If problems crop up, we can add back approximation tests, but for now, I
think it's more important to have a clear path for the cleanups I'm
working on.