Merge `SweptCurve` into `Surface`
This is a simplification that I should have done a long time ago.

I figured, it is only a matter of time, before we add more types of
surfaces, so I thought it made sense to keep `Surface` an enum. That
might well be true, but in the meantime, this decision makes lots of
code more complicated. Once we want to add more types of curves, we can
easily turn `Surface` into an enum again.