Invert zoom direction
The intent behind the original zoom direction was to provide a
consistent metaphor: You manipulate the model, not an abstract camera.

Unfortunately, it turned out to not be that simple. At least on Gnome,
if you enable "natural scrolling" (which is the same metaphor, applied
to the Gnome UI), you invert all mouse whell input, and so also the zoom
direction in Fornjot.

With this change, the zoom direction in Fornjot follows the "natural
scrolling" configuration in Gnome. Unfortunately I don't have data from
other platforms.