Don't use `actions-rs` to set up toolchain
We have a `rust-toolchain.toml` in the repository, and `rustup` is
installed on GitHub Actions, as far as I know. I think this should just
do the right thing.

We might lose caching with this, but I'm not sure that's worth it. Maybe
once there is a maintained `actions-rs` alternative.