Remove uniqueness validation
It is the only check left that runs on the old validation
infrastructure, and I'm not sure it actually does something useful
anymore, given the coverage provided by all the other checks. If it
turns out it does, we can re-add some form of it later, based on the new
validation infrastructure.