Move `path` to new `geometry` module
I'm about to add new types that are similar to the path types, and
having them in a common parent module makes sense. `geometry` isn't the
greatest name for that new module (it's not like the things outside of
it aren't geometry), but it's what I can come up with right now.