Fix potential soundness hole
Store `Version` in the version `static`s, instead of storing a `&'static
str` there and converting the version from that.

Going over the relevant parts of the standard library documentation with
a fine comb, I realized that a pointer casted from a reference is only
valid, if only raw pointers are used to access that memory.[1]

I don't know if this carries over to the pointer returned by
`str::as_ptr`, but if it does, this could have indeed been a soundness
hole, as the original `str` is still used, in `fj-app`'s `args` module.

[1]: https://doc.rust-lang.org/std/ptr/index.html#safety