Simplify `PartialHalfEdge::build`
This commit simplifies one method, at the expense of making a whole lot
of other code more complicated. Simplifying `build` helps with the next
step of replacing the old partial object API. Once more of the new API
is in place, it should be possible to simplify the new complicated code
again.