Remove unused builder methods
The unit tests that used these methods no longer exist, and as I found
out, that code was very dodgy to begin with. It worked in the contexts
where it was used, but it probably wouldn't have worked in real code, as
it didn't correctly take care of face orientation.

Pretty sure those cubes wouldn't have made it through the 3MF validation
testing.