Fix wgpu validation error
wgpu would fail after the window had been minimized and was un-minimized
again after a while, because `Gui::update` would run many times while
minimized, but `Gui::prepare_draw` wouldn't. That would lead to
`end_frame` not being called for a long time, which in turn would lead
to egui accumulating output.

After that went on for a while, and `prepare_draw` was called again,
there would be too much output for the index buffer to handle.