Remove sweep tests
Those tests are a constant thorn in my side. They resist every
refactoring I do, and always need to be carefully massaged to keep
working. I'm facing another of those moments, and I have enough.

The results of the sweep are somewhat easy to inspect manually, and in
combination with validation checks and export validation, it should be
easy enough to keep the sweep code under control, despite the lack of
direct automated testing.

This is not ideal, of course. But I need to make progress. And since the
sweep code will likely not survive the coming clean-ups in its current
form anyway, it probably makes more sense to build up a new test suite
in parallel to whatever the rewritten sweep code ends up being.