Attempt at shell validation
Attempt to validate `Shell`'s by making sure that each half-edge in a
shell refers to a global-edge that is referred to by another half-edge
in that same shell. My logic says that each edge in a shell should be a
part of exactly two faces but I may be wrong.

This validation check currently fails with `cargo run --package export-
validator`, but I'm not sure if this is a bug in my validation checks,
or whether I exposed a bug in existing code.

For example looking at the cuboid example, many global edges are
referred to by only one half_edge, which seems wrong to me...