Don't store service events
Eventually, this will be needed to display history in the app. But for
now, it's not being used, and it's getting in the way of other changes
I'm making.

It can be re-added later, as needed, and then, it'll hopefully be
clearer how to better integrate it with the rest of the design.