Remove horrible hack
So, as the comment suspected, this did come back to bite me! Fortunately
not in a bad way, as I found this comment pretty quick while figuring
out what was wrong with the new code I wrote. But honestly, I'm
surprised that it lasted this long (almost 10 months!).

To my delight, removing this hack required no changes to any code! I
take that as a sign that we're in a much better position today,
architecture-wise, then we were back then.