Fix curve cache insertion bug
Curve insertion *returned* the correct curve segment, which means either
the existing segment, if there is one, or the new one being inserted.
But it actually *inserted* the wrong one, i.e. the new one.

I don't think this lead to any bugs, due to the current usage patterns
of the cache API, but it's wrong none the less, and a bug waiting to
happen.