Rename `Edge` to `HalfEdge`
This reverts a decision made some time ago, after `GlobalEdge` had been
removed and the name `Edge` became free.

While `Edge` is shorter and thus nicer, it's also less clear. I found
that it's easy to forget that this object is not actually "the edge",
but only, well, a half-edge that always had an equal but opposite
sibling.

And that's despite me working on this almost daily, so I guess it's
potentially even more confusing for someone less active in the code
base.