Don't insert updated replacement result
This makes replace operations more flexible, as it gives control over
insertion to the caller. This is especially relevant, if the replacement
produces an invalid intermediate step. Currently, everything that is
inserted is validated, and validation provides no good mechanism to deal
with invalid intermediate objects, even if the final object ends up
being valid.