Allow valid `Cycle`s to be empty
That empty cycles are invalid makes the result of `BuildFace::unbound`
invalid by default, requiring a hacky workaround
(`Services::only_validate`).

I think empty cycles are actually fine. What's a problem is, if such an
empty cycle is used as the exterior of a face, and I will add a
validation check for that specifically.