Implement `ValidationCheck` for the check
The way it was previously, fully qualified method call syntax would have
been required to run a validation check. This wasn't visible in the code
so far, as there only is one validation check currently.

This change makes running validation checks a bit more convenient
(compared to what it would have needed to be, not what's currently in
the code). It preserves the discoverability, as the implementations are
still listed in the documentation of the object type.