Remove redundant validation check
This check essentially verified that the length of a half-edge is
non-zero, but I don't think this is actually important. At least I can't
think of a reason why it wouldn't work, and there might be legitimate
use cases.

What is important, is whether the vertices that bound the half-edge are
coincident, but in this regard, this check could have lead to false
positives, as the half-edge could be bounded by the same vertex on both
sides.

In addition, there are other validation checks that check for coincident
vertices, either existing in code, or tracked in issues.