Make `PathApproxParams` public
I need to use it from the `geometry` module. I expect that long-term,
`approx` and `geometry` will get merged, and this code will turn into an
internal implementation detail again.

But for now, I don't want to rearchitect the whole module structure,
before I have a better grasp on what the new geometry system is going to
look like. So just making it public is the more pragmatic choice.