Remove curve geometry mismatch validation check
The CI build is currently failing because of that check. This has
started happening due to a change of how circles are represented in the
new geometry system. Parts of the new geometry system are being used by
this check,

The cause of the problem seems to be that the curves that are failing
the check have their geometry defined as a circle on a straight surface,
and a line on a curved surface. The new geometry system, which is
designed to eliminate this kind of redundancy, doesn't deal with that
well, and delivers slightly different results for each case.

This doesn't actually matter though. Only one of those definitions is
actually used. When disabling the check, the affected models look fine
in the viewer, and most importantly, they validate just fine when
exported to 3MF.

I don't think this validation check was ever there to prevent invalid
geometry. Its purpose was to discover bugs, which these kinds of
discrepancies are a sign of. But here, it wasn't detecting a bug (as far
as I can tell, at least). It was detecting a discrepancy due to the
ongoing transition to the new geometry system.

The new geometry system will eliminate the redundancies that this check
was designed to keep in check. It would have become obsolete anyway. In
light of this, I've decided it has already outlived its purpose.

The alternative would be to somehow paper over this discrepancy until
the redundancy can be eliminated. But I don't know how to do that. I
suspect it would have been hard. And I've already spent enough time
tracking this problem down.

I think it's fine taking the easy way out, and retiring this validation
check a bit early.