Move `Circle` back to `fj-match`
For a while, it seemed to make sense to consolidate all curve-related
code in `fj-core`. But I think that was a mistake.

At least `Line` would need to move back, because I'm about to add a
dependency on that from `LineSegment`, which lives in `fj-math`. I could
move `LineSegment` to `fj-core`, of course, but then I'd basically need
to add a `math` module there, and I'm don't think that makes sense.

Anyway, if `Line` has to move back, it would be inconsistent to leave
just `Circle`.