No longer render vertices
Vertices are currently exposed through `Operation` in a privileged way,
and with the new topographical operations incoming, this no longer makes
sense.

Rendering vertices depends on this privileged way. I'm removing it, so I
can then clean up `Operation` itself. Right now, this feature isn't
adding a lot of value, so I don't think it's much of a loss.

It seems desirable to render vertices again at some point. But it also
seems desirable to render other things that currently aren't, like
planes (or surfaces in general). It makes sense that we'd grow a
mechanism for things with specialized rendering in the future, but right
now I neither know what that would be, nor do I want to be weighed down
by it.