Display `Shape` children by themselves
Being able to go through the children of shape, seeing how each is added
one by one, was kinda neat, but not that useful in the end. Being able
to see each child in isolation instead, is about as useful and quite a
bit simpler.

And there is one way, in which the previous way started getting in the
way: When building up the shape child by child, sometimes the
construction of some children could not be observed, as other children
had already been constructed in front. This is the reason that I decided
to abandon the previous approach.