Add copy of `fj-viewer`
I'd like to expand the scope of the experimentation to include the
visualization again. I think having a copy of `fj-viewer` is the most
practical way to do that. If all goes well, any changes I make here,
will be merged back into `fj-viewer`, once the experimental code overall
is merged back into the mainline code.